updated readme and moved authentication code to seperate file
